+++ common/src/main/java/mca/entity/ai/goal/GrimReaperRestGoal.java
@@ -12,7 +12,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 
 public class GrimReaperRestGoal extends Goal {
@@ -20,6 +19,8 @@ public class GrimReaperRestGoal extends Goal {
     private final GrimReaperEntity reaper;
     private int lastHeal = -COOLDOWN;
     private int healingCount = 0;
+    private static final int MAX_HEALING_COUNT = 5;
+    private static final int MAX_HEALING_TIME = 500;
     private int healingTime;
 
     public GrimReaperRestGoal(GrimReaperEntity reaper) {
@@ -28,7 +29,7 @@ public class GrimReaperRestGoal extends Goal {
 
     @Override
     public boolean canStart() {
-        return reaper.age > lastHeal + COOLDOWN && reaper.getHealth() <= (reaper.getMaxHealth() / (healingCount + 1.5));
+        return reaper.age > lastHeal + COOLDOWN && reaper.getHealth() <= (reaper.getMaxHealth() * (1.0f - (healingCount + 1.0f) / (float)MAX_HEALING_COUNT));
     }
 
     @Override
@@ -45,7 +46,7 @@ public class GrimReaperRestGoal extends Goal {
     public void start() {
         reaper.requestTeleport(reaper.getX(), reaper.getY() + 8, reaper.getZ());
 
-        healingTime = 500;
+        healingTime = MAX_HEALING_TIME;
         lastHeal = reaper.age;
         healingCount++;
     }
@@ -62,24 +63,35 @@ public class GrimReaperRestGoal extends Goal {
 
         reaper.setVelocity(Vec3d.ZERO);
 
-        if (!reaper.world.isClient && healingTime % 10 == 1) {
-            reaper.setHealth(reaper.getHealth() + MathHelper.clamp(1.0F / healingCount, 0.5F, 1.0F));
+        if (!reaper.world.isClient && healingTime % (10 + healingCount * 5) == 0) {
+            reaper.setHealth(reaper.getHealth() + 1);
         }
 
-        if (!reaper.world.isClient && healingTime % 100 == 1) {
+        if (!reaper.world.isClient && healingTime % 50 == 0) {
             // Let's have a light show.
-            int dX = reaper.getRandom().nextInt(8) + 4 * reaper.getRandom().nextFloat() >= 0.50F ? 1 : -1;
-            int dZ = reaper.getRandom().nextInt(8) + 4 * reaper.getRandom().nextFloat() >= 0.50F ? 1 : -1;
-            int y = TaskUtils.getSpawnSafeTopLevel(reaper.world, (int) reaper.getX() + dX, 256, (int) reaper.getZ() + dZ);
+            int dX = reaper.getRandom().nextInt(16) - 8;
+            int dZ = reaper.getRandom().nextInt(16) - 8;
+            int y = TaskUtils.getSpawnSafeTopLevel(reaper.world, (int)reaper.getX() + dX, 256, (int)reaper.getZ() + dZ);
 
-            EntityType.LIGHTNING_BOLT.spawn((ServerWorld) reaper.world, null, null, null, new BlockPos(dX, y, dZ), SpawnReason.TRIGGERED, false, false);
+            EntityType.LIGHTNING_BOLT.spawn((ServerWorld)reaper.world, null, null, null, new BlockPos(reaper.getX() + dX, y, reaper.getZ() + dZ), SpawnReason.TRIGGERED, false, false);
 
-            // Also spawn a random skeleton or zombie.
-            EntityType<?> m = reaper.getRandom().nextFloat() >= 0.5F ? EntityType.ZOMBIE : EntityType.SKELETON;
-            Entity e = m.spawn((ServerWorld) reaper.world, null, null, null, new BlockPos(reaper.getX() + dX + 4, y, reaper.getZ() + dZ + 4), SpawnReason.TRIGGERED, false, false);
+            if (!reaper.world.isClient && healingTime % 100 == 0) {
+                // Also spawn a random enemy
+                EntityType<?> m = reaper.getRandom().nextFloat() < 0.5f ? EntityType.ZOMBIE : EntityType.SKELETON;
+                Entity e = m.spawn((ServerWorld)reaper.world, null, null, null, new BlockPos(reaper.getX() + dX, y, reaper.getZ() + dZ), SpawnReason.TRIGGERED, false, false);
 
-            if (e != null && m == EntityType.SKELETON) {
-                e.equipStack(EquipmentSlot.MAINHAND, new ItemStack(Items.BOW));
+                // Equip them
+                if (e != null) {
+                    if (m == EntityType.SKELETON) {
+                        e.equipStack(EquipmentSlot.MAINHAND, new ItemStack(Items.BOW));
+                    } else {
+                        e.equipStack(EquipmentSlot.MAINHAND, new ItemStack(Items.IRON_SWORD));
+                    }
+                    e.equipStack(EquipmentSlot.HEAD, new ItemStack(Items.IRON_HELMET));
+                    e.equipStack(EquipmentSlot.CHEST, new ItemStack(Items.IRON_CHESTPLATE));
+                    e.equipStack(EquipmentSlot.LEGS, new ItemStack(Items.IRON_LEGGINGS));
+                    e.equipStack(EquipmentSlot.FEET, new ItemStack(Items.IRON_BOOTS));
+                }
             }
         }
     }