+++ changelog.md
@@ -190,3 +190,4 @@
 * Buildings now support modded chests
 * Villagers will now use your editor name
 * Fixed letter author and creative mode usage
+* Strengthened Grim Reaper